+++ source/chip/STM32/STM32F0/Tupfile.lua
@@ -19,6 +19,8 @@ if CONFIG_CHIP_STM32F0 == "y" then
 			"\" > \"%o\"", ldscriptOutputs)
 
 	CXXFLAGS += STANDARD_INCLUDES
+	CXXFLAGS += ARCHITECTURE_INCLUDES
+	CXXFLAGS += CHIP_INCLUDES
 
 	tup.include(DISTORTOS_TOP .. "compile.lua")